+++ rest_framework_json_api/relations.py
@@ -116,7 +116,19 @@ class HyperlinkedMixin(object):
         })
         self_link = self.get_url('self', self.self_link_view_name, self_kwargs, request)
 
-        related_kwargs = {self.related_link_url_kwarg: kwargs[self.related_link_lookup_field]}
+        """
+        Assuming RelatedField will be declared in two ways:
+        1. url(r'^authors/(?P<pk>[^/.]+)/(?P<related_field>\w+)/$',
+                AuthorViewSet.as_view({'get': 'retrieve_related'}))
+        2. url(r'^authors/(?P<author_pk>[^/.]+)/bio/$',
+                AuthorBioViewSet.as_view({'get': 'retrieve'}))
+        So, if related_link_url_kwarg == 'pk' it will add 'related_field' parameter to reverse()
+        """
+        if self.related_link_url_kwarg == 'pk':
+            related_kwargs = self_kwargs
+        else:
+            related_kwargs = {self.related_link_url_kwarg: kwargs[self.related_link_lookup_field]}
+
         related_link = self.get_url('related', self.related_link_view_name, related_kwargs, request)
 
         if self_link: